requirements.txt
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-UserNotifications>=9.0
